[function-signature-opts] Split SignatureAnalyzer into a separate function analysis pass.
This changes moves SignatureAnalyzer and ArgumentDescriptor/ResultDescriptor into a separate
pass.

This pass is run on every function and the optimized signature is return'ed through the
getArgDescList and getResultDescList.

Next step is to split to cloning and callsite rewriting into their own function passes.
