+++ include/swift/SILOptimizer/Analysis/Analysis.def
@@ -29,6 +29,7 @@ ANALYSIS(ClassHierarchy)
 ANALYSIS(Destructor)
 ANALYSIS(Dominance)
 ANALYSIS(Escape)
+ANALYSIS(FunctionSignature)
 ANALYSIS(InductionVariable)
 ANALYSIS(Loop)
 ANALYSIS(LoopRegion)
